LayoutTests/platform/mac/imported/w3c/web-platform-tests/uievents/keyboard/modifier-keys-combinations-expected.txt

Target

PASS Check sending "Shift + y" key combination
PASS Check sending "Control + y" key combination
PASS Check sending "Alt + y" key combination
FAIL Check sending "Meta + y" key combination assert_equals: expected "y" but got "leftMeta"
